requirements.txt
blinker==1.4
click==8.0.1
Flask==2.0.1
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.8.6
py-bcrypt==0.4
SQLAlchemy==1.4.18
Werkzeug==2.0.1
